% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{address2path}
\alias{address2path}
\alias{path2address_sufix}
\alias{path2address}
\alias{path_tree}
\title{Address}
\usage{
address2path(x, address = "file://", path = "")

path2address_sufix(x, address = "file://")

path2address(x, address = "file://")

path_tree(path, hide = FALSE)
}
\arguments{
\item{x}{Path.}

\item{address}{Server address, Default: 'file://'.}

\item{path}{Path equivalent of the address, Default: ''.}

\item{hide}{Make it collapsible, Default: FALSE.}
}
\value{
Path.

Path ready to go after 'address'.

Address.

Prints the tree.
}
\description{
This transforms an address in the server to a path.

Transform a path to a server address sufix (after object@address)

Takes a path and transforms it into and address.

Takes a path and transforms it into a tree.
}
\examples{
\dontrun{
 if(interactive()){
  address2path("file://Groups/mypath/step")
  address2path(
    x = "https://my_website/Groups/mypath/step",
    address = "https://my_website/Groups",
    path = "/mnt/Groups"
  )
 }
}
\dontrun{
 if(interactive()) path2address_sufix("/mnt/Groups/mypath/step")
}
\dontrun{
 if(interactive()) path2address("/mnt/Groups/mypath/step")
}
\dontrun{
 if(interactive()) path_tree("~/Documents", hide = TRUE)
}
}
