% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_wsteps}
\alias{create_wsteps}
\title{Create Reportr object}
\usage{
create_wsteps(
  steps,
  project.name = "My project",
  author = Sys.info()[["user"]],
  affiliation = "Lab",
  email = paste0(Sys.info()[["user"]], "@email.com"),
  address = "file://",
  path = "",
  markdown = list("")
)
}
\arguments{
\item{steps}{List of paths (and other arguments embedded in each element
of the list).}

\item{project.name}{The name of your workflow/project, Default = 'My project'.}

\item{author}{Author, Default = Sys.info user.}

\item{affiliation}{Author, Default = 'Lab'.}

\item{email}{Author's email.}

\item{address}{Server address, Default: 'file://'.}

\item{path}{Path equivalent of the address, Default: ''.}

\item{markdown}{Markdown specs.}
}
\value{
A Reportr object.
}
\description{
\code{create_wsteps} creates the Reportr object and the necessary
structure for the package's functions.
}
\examples{
\dontrun{
 if(interactive()) create_wsteps(list())
}
}
