% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Summarise table}
\usage{
get_summary_table(mdata, min_n = 1)
}
\arguments{
\item{mdata}{Object of class data.frame.}

\item{min_n}{Filter columns with this number of objects or less, Default: 1.}
}
\value{
data.frame.
}
\description{
\code{get_summary_table} summarise each column of a table.
}
\details{
It only takes the columns of class character or factor.
}
\examples{
\dontrun{
 if(interactive()) get_summary_table(reshape2::melt(Titanic))
}
}
\seealso{
\code{\link[reshape2]{melt}}
}
