% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{highlight_f_def}
\alias{highlight_f_def}
\alias{report_files}
\alias{highlight_f}
\title{File definition}
\usage{
highlight_f_def(
  x,
  title_pattern = c("pdf$"),
  path = "no_path/found",
  lock = FALSE
)

report_files(confs, name, path = "/no/path", lock = FALSE, ...)

highlight_f(
  hfiles,
  path,
  hlevel = "###",
  close_tabset = FALSE,
  address = "file://",
  verbose = FALSE
)
}
\arguments{
\item{x}{Step config or just the path to the step's output.}

\item{title_pattern}{File \code{regex}, Default: c("pdf$")}

\item{path}{Step's output path.}

\item{lock}{Whethe to look for 'locking plots in clustering', Default: FALSE.}

\item{confs}{List of config steps.}

\item{name}{Name in the 'confs' list.}

\item{...}{More parameters for 'highlight_f'.}

\item{hfiles}{\code{regex} to find the file.}

\item{hlevel}{Header level, Default: '###'.}

\item{close_tabset}{In case you need something to show right after the
last tab, Default: FALSE.}

\item{address}{Server address, Default: 'file://'.}

\item{verbose}{Show progress, Default: FALSE.}
}
\value{
Config list with the highlighting \code{regex} patterns.

All lines created are showed with 'cat'.

Adds the file to the report.
}
\description{
Define the \code{regex} pattern to the files to highlight.

It finds plots given a vector of regex paths/

Add highlighted files (mainly tables and Graphics).
}
\details{
This might not work without {r, result='asis'}.

Tries to find the file using the \code{regex}.
}
\examples{
\dontrun{
 if(interactive()) highlight_f_def("~/Documents/liai/literature")
}
\dontrun{
 if(interactive()){
   report_files(
     confs = list(demx = list("/path/2/output", hfiles = 'pdf$')),
     name = "demx", path = "/path/2/output"
   )
 }
}
\dontrun{
 if(interactive()) highlight_f("pdf", "~/Documents/liai/literature")
}
}
\seealso{
\code{\link[stringr]{case}}
\code{\link[data.table]{fread}}
\code{\link[knitr]{kable}}
}
