% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{make_diagram}
\alias{make_diagram}
\title{Create workflow diagram}
\usage{
make_diagram(
  report_file = knitr::current_input(dir = TRUE),
  critical = list(names = "lustering|uality", color = "Gold"),
  exclude = "libraries",
  style = "single-cell",
  extra_connections = NULL,
  critical_connections = NULL,
  central_node = "",
  last_node = ""
)
}
\arguments{
\item{report_file}{Step's list or Reportr object,
Default: knitr::current_input(dir = TRUE).}

\item{critical}{Critical steps,
Default: list(names = "lustering|uality", color = "Gold").}

\item{exclude}{Exclude these steps, Default: 'libraries'.}

\item{style}{Style has a pre-set of connections between steps,
Default: 'single-cell'.}

\item{extra_connections}{This further modifies connections, Default: NULL.}

\item{critical_connections}{Makes the connection red, Default: NULL.}

\item{central_node}{Step connected to the rest after it, Default: ''.}

\item{last_node}{Last step in the diagram, Default: ''.}
}
\value{
Spec for a Graphviz diagram as text.
}
\description{
\code{make_diagram} creates a
}
\details{
At the moment is more difficult to get the right titles from a list or
an Reportr object
}
\examples{
\dontrun{
 if(interactive()){
  DiagrammeR::grViz(make_diagram(list("Step1"= 1, "Step2" = 2), style = NULL))
 }
}
}
\seealso{
\code{\link[knitr]{current_input}}
}
