% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_title}
\alias{make_title}
\title{Title format}
\usage{
make_title(x)
}
\arguments{
\item{x}{String to format.}
}
\value{
Adapted string.
}
\description{
Create title format.
}
\examples{
\dontrun{
 if(interactive()) make_title("the title of my famous book")
}
}
\seealso{
\code{\link[stringr]{case}}
}
