% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{markdown_check}
\alias{markdown_check}
\title{Markdown header}
\usage{
markdown_check(object, verbose = FALSE)
}
\arguments{
\item{object}{Reportr object.}

\item{verbose}{Show progress, Default: FALSE.}
}
\value{
Text representation of an R markdown file's header.
}
\description{
Check all necessary section of the YAML header are present.
}
\details{
If sections are not in the YAML, adds them from the object.
}
\examples{
\dontrun{
 if(interactive()){
  cat(markdown_check(object = create_wsteps(list())))
  cat(markdown_check(object = create_wsteps(list(),
    markdown = list("~/Documents/liai/scripts/reportr_pkg/reportr_webspec.yaml")
  )))
  cat(markdown_check(object = create_wsteps(list(),
    markdown = list(output=list(html_document=list(theme='journal')))
  )))
 }
}
}
\seealso{
\code{\link[yaml]{read_yaml}}, \code{\link[yaml]{as.yaml}}
}
