% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_step}
\alias{process_step}
\title{Step output}
\usage{
process_step(
  object,
  name = NULL,
  address = "file://",
  path = "/mnt/BioAdHoc/Groups/"
)
}
\arguments{
\item{object}{Reportr object.}

\item{name}{Name of the step to process, Default: NULL.}

\item{address}{Server address, Default: 'file://'.}

\item{path}{Path equivalent of the address, Default: '/mnt/BioAdHoc/Groups/'.}
}
\value{
List with output path and config (the rest of the information in the)
the step element).
}
\description{
Finding step output path.
}
\details{
It can also take a yaml file with output_dir(/project).
}
\examples{
\dontrun{
 if(interactive()){
  process_step("~")
 }
}
}
\seealso{
\code{\link[yaml]{read_yaml}}
}
