% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{report_link}
\alias{report_link}
\title{Link output}
\usage{
report_link(x, address = "file://")
}
\arguments{
\item{x}{Path to step's output.}

\item{address}{Server address, Default: 'file://'.}
}
\value{
Prints markdown link.
}
\description{
Create link to output path.
}
\examples{
\dontrun{
 if(interactive()) report_link("~/Documents/liai/literature")
}
}
\seealso{
\code{\link[stringr]{case}}
}
