% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{report_render}
\alias{report_render}
\title{Render Reportr object}
\usage{
report_render(object, rmd_file = NULL, ...)
}
\arguments{
\item{object}{Reportr object.}

\item{rmd_file}{Preserve the generated Rmd, Default: NULL.}

\item{...}{Extra parameters for render from rmarkdown.}
}
\value{
NULL. It renders the object.
}
\description{
\code{report_render} parses the Reportr object, creates a temporary
Rmd file and then calls render from markdown to create the report file.
}
\examples{
\dontrun{
 if(interactive()){
  report_render(create_wsteps(as.list(list.dirs("~/Documents", recursive = FALSE))))
 }
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
