% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{report_set_titles}
\alias{report_set_titles}
\title{Create step titles}
\usage{
report_set_titles(object)
}
\arguments{
\item{object}{Reportr object.}
}
\value{
Reportr object with titles in each step.
}
\description{
\code{report_set_titles} takes the list of steps and uses the names
as titles.
}
\examples{
\dontrun{
 if(interactive()) report_set_titles(create_wsteps(list(`step n`="path")))@steps
}
}
