% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{report_steps}
\alias{report_steps}
\title{Step's report}
\usage{
report_steps(object)
}
\arguments{
\item{object}{Reportr object.}
}
\value{
Rendered Reportr object into r markdown.
}
\description{
\code{report_steps} goes over the steps of a Reportr object and.
deployes their elements.
}
\examples{
\dontrun{
 if(interactive()) report_steps(create_wsteps(list(`step n`="path")))
}
}
