% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{report_yaml}
\alias{report_yaml}
\title{Report from YAML}
\usage{
report_yaml(file, rmd_file = NULL, return_object = FALSE, ...)
}
\arguments{
\item{file}{Path to the file.}

\item{rmd_file}{Preserve the generated Rmd, Default: NULL.}

\item{return_object}{Whether or not to retrun the object, Default: FALSE.}

\item{...}{Extra parameters for render from rmarkdown.}
}
\description{
Takes a YAML file with steps as a dictionary/list.
}
\details{
A Reportr object saved in a temporary file as wells as the Rmd.
}
\examples{
\dontrun{
 if(interactive()) report_yaml('vignette/example1.yaml')
}
}
\seealso{
\code{\link[yaml]{read_yaml}}
}
