% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Reportr-class}
\alias{Reportr-class}
\title{An S4 class to represent your workflow.}
\description{
An S4 class to represent your workflow.
}
\section{Slots}{

\describe{
\item{\code{steps}}{A list of each step output in your workflow.}

\item{\code{project.name}}{The tilte of your project.}

\item{\code{author}}{Person creating the report.}

\item{\code{affiliation}}{Institute, Department, or University, etc.}

\item{\code{email}}{Author's email.}

\item{\code{address}}{Server address, Default: 'file://'.}

\item{\code{path}}{Path (in disk) equivalent of the address, Default: ''.}
}}

