% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{report_title}
\alias{report_title}
\alias{report_filters}
\alias{report_parameters}
\alias{report_tables}
\alias{report_description}
\alias{report_gsheet}
\alias{report_issues}
\alias{report_command}
\alias{report_section}
\title{Step title}
\usage{
report_title(x, conf, hlevel = "###")

report_filters(x)

report_parameters(x)

report_tables(x, interactive = "choose", table_summarise = FALSE)

report_description(x, conf = NULL, hide = FALSE)

report_gsheet(x)

report_issues(x, name = "", hlevel = "###")

report_command(x, type = "bash")

report_section(
  object,
  name = 1,
  hide = FALSE,
  lock = FALSE,
  hlevel = "###",
  close_tabset = FALSE
)
}
\arguments{
\item{x}{Step's config.}

\item{conf}{Step's arguments, Default: NULL.}

\item{hlevel}{Header level, Default: '###'.}

\item{interactive}{Use DT?, Default: 'choose' (if No. of rows > 15).}

\item{table_summarise}{Make a summary of the table, Default: FALSE.}

\item{hide}{Collapsible description, Default: FALSE.}

\item{name}{Step's name, Default: 1.}

\item{type}{Type of chunk (r, bash, python).}

\item{object}{Reportr object.}

\item{lock}{Is it a step with "locked" parameters, Default: FALSE.}

\item{close_tabset}{In case you need something to show right after the
last tab, Default: FALSE.}
}
\value{
Prints the title using \code{cat}.

None (invisible 'NULL'). If one filter it will show a line, if not
it will use 'knitr::kable' after reshape2::melt.

Prints a kable.

Creates a table.

Prints the description.

Prints an iframe object.

Prints the issues (if found).

Prints the command into a selected chunk type.

Prints all the elements of a step.
}
\description{
Show a title given to the step.

Add a line or table of used filters in a given step. Since a
step should probably come from a YAML file, then it'll iteratively look for
filtering, filters, filter, in that order. Lastly, it looks for subset.

Show parameters used for a step.

Show tables with DT::datatable or knitr::kable.

Add description from a step element.

Connect google document.

Show step's issues.

Show step's command(s).

Create step's section.
}
\details{
It accepts a single filtering criteria or multiple (melted).

It accepts a list, looks for "lock" or "parameters" and takes the
first one it finds.

Creates a temporary Rmd file and calls it with \code{knit_child}.
}
\examples{
\dontrun{
 if(interactive()) {
  report_title(x = "Step N", conf = list("~/Documents/liai/literature"))
 }
}
\dontrun{
  if(interactive()){
    report_filters(x = list(filtering = "n_umi >= 1000"))
  }
}
\dontrun{
  if(interactive()) report_parameters(list(parameters = NULL))
}
\dontrun{
 if(interactive()) {
  report_tables(x = "~/Documents/liai/asthma/biopsy/metadata_library.csv")
 }
}
\dontrun{
 if(interactive()){
  report_description("Description")
  report_description(list(description = "Brief description"))
  report_description("Description", hide = TRUE)
 }
}
\dontrun{
 if(interactive()) {
  report_gsheet("spreadsheets/d/1v3ff8KKWGK5iCAyMa0v45wiw5El4Lcy0HVjTAbuYguU")
 }
}
\dontrun{
 if(interactive()) {
  report_issues(list(issues = "Red is not a colour"))
  report_issues(list(issues = "Solved: red _is_ a colour"))
  report_issues(list(issues = "Fatal: red _is_ blue!"))
 }
}
\dontrun{
 if(interactive()) report_command(list(command = "sh /path/to/script.sh"))
}
\dontrun{
 if(interactive()) {
  mysteps = list(s1 = list("~/Documents/liai/literature", issues = "Oh no!"))
  mysteps$s1$hfiles = c(`Cool review` = "immunology.pdf")
  ws = create_wsteps(mysteps)
  report_section(ws)
 }
}
}
\seealso{
\code{\link[reshape2]{melt}}
\code{\link[knitr]{kable}}

\code{\link[knitr]{kable}}

\code{\link[data.table]{fread}}, \code{\link[data.table]{rbindlist}}
\code{\link[DT]{datatable}}
\code{\link[knitr]{kable}}

\code{\link[knitr]{knit_child}}
}
