% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wreportr-data}
\alias{wreportr-data}
\alias{highlight_defaults}
\alias{preset_titles}
\alias{preset_description}
\title{Useful variables}
\format{
Character lists.
}
\usage{
highlight_defaults

preset_titles

preset_description
}
\description{
These variables will help us document each of the steps. They currently
tailored to single-cell analysis, but can be easily extended.
}
\examples{
\dontrun{
 length(highlight_defaults)
 highlight_defaults[1:5]

 length(preset_titles)
 preset_titles[1:5]

 length(preset_description)
 preset_description[1:5]
}
}
\keyword{datasets}
