% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportr.R
\docType{package}
\name{wreportr}
\alias{wreportr}
\title{wreportr: A package for easily generating reports.}
\description{
The wreportr package provides 3 categories of important functions:
report, diagram, and highlight.
}
\section{The Reportr class}{

It helps just put together the steps of your workflow. Its structure is:
Steps, project name, author, affiliation, email, address (server), and
path (disk).
}

\section{Report functions}{

These are designed to help you generating the sections of a report.
Title, description, command, filters, parameters, tables, issues,
link to results, and highlighted files.
}

